refactor: extract shared console route helpers, fix last Sonar issue
Extract httpStatusForStoreError, sendStoreError, and
getNormalizedStringField into consoleRouteHelpers.ts — eliminates the
19-line duplicated block between totpRoutes.ts and tokenRoutes.ts that
Sonar flagged as 2.6% duplication on new code.

Also fixes the remaining S7780 (String.raw for backslash escaping) in
VerificationNotifier.ts — both occurrences on lines 100-101 now use
String.raw template literals consistently.
