+++ src/web/routes/tokenRoutes.ts
@@ -18,10 +18,10 @@
 
 import express, { Router } from 'express';
 import type { Request, Response } from 'express';
-import { TotpError, type ConsoleTokenStore, type TotpErrorCode } from '../console/consoleToken.js';
+import { TotpError, type ConsoleTokenStore } from '../console/consoleToken.js';
 import { createAuthMiddleware } from '../middleware/authMiddleware.js';
 import { SlidingWindowRateLimiter } from '../../utils/SlidingWindowRateLimiter.js';
-import { UnicodeValidator } from '../../security/validators/unicodeValidator.js';
+import { httpStatusForStoreError, sendStoreError, getNormalizedStringField } from './consoleRouteHelpers.js';
 import { logger } from '../../utils/logger.js';
 
 /** JSON body size limit — rotation requests are tiny. */
@@ -34,57 +34,6 @@ const BODY_LIMIT = '1kb';
  */
 const DEFAULT_RATE_LIMIT_MAX = 10;
 const DEFAULT_RATE_LIMIT_WINDOW_MS = 60_000;
-
-/**
- * Maps a store error code to an HTTP status. Replicates the pattern from
- * totpRoutes.ts so both route files return consistent status codes for the
- * same failure class.
- */
-function httpStatusForTokenError(code: TotpErrorCode): number {
-  switch (code) {
-    case 'ALREADY_ENROLLED':
-      return 409;
-    case 'NOT_ENROLLED':
-    case 'PENDING_NOT_FOUND':
-    case 'INVALID_TOTP_CODE':
-      return 400;
-    case 'TOO_MANY_PENDING':
-      return 429;
-    case 'TOTP_REQUIRED':
-      return 403;
-    case 'STORE_NOT_INITIALIZED':
-      return 503;
-    default: {
-      code satisfies never;
-      return 400;
-    }
-  }
-}
-
-/**
- * Send a structured error response with both a human-readable message and
- * a machine-readable code. Same shape as the TOTP routes.
- */
-function sendTokenError(
-  res: Response,
-  status: number,
-  code: string,
-  message: string,
-): void {
-  res.status(status).json({ error: message, code });
-}
-
-/**
- * Safely extract a string field from an unknown request body and NFC-normalize
- * it (DMCP-SEC-004). Returns null if the body is not an object or the field
- * is missing / not a string.
- */
-function getNormalizedStringField(body: unknown, field: string): string | null {
-  if (!body || typeof body !== 'object') return null;
-  const val = (body as Record<string, unknown>)[field];
-  if (typeof val !== 'string') return null;
-  return UnicodeValidator.normalize(val).normalizedContent;
-}
 
 /**
  * Options for the token routes factory.
@@ -122,12 +71,12 @@ export function createTokenRoutes(options: TokenRoutesOptions): Router {
   /** POST /rotate — rotate the primary console token with TOTP confirmation. */
   router.post('/rotate', jsonParser, async (req: Request, res: Response) => {
     if (!rotateLimiter.tryAcquire()) {
-      sendTokenError(res, 429, 'RATE_LIMITED', 'Too many rotation attempts — slow down');
+      sendStoreError(res, 429, 'RATE_LIMITED', 'Too many rotation attempts — slow down');
       return;
     }
     const confirmationCode = getNormalizedStringField(req.body, 'confirmationCode');
     if (!confirmationCode) {
-      sendTokenError(res, 400, 'MISSING_FIELDS', 'confirmationCode is required');
+      sendStoreError(res, 400, 'MISSING_FIELDS', 'confirmationCode is required');
       return;
     }
     try {
@@ -141,9 +90,9 @@ export function createTokenRoutes(options: TokenRoutesOptions): Router {
       const message = err instanceof Error ? err.message : 'Rotation failed';
       logger.debug(`[Token] rotate failed: ${message}`);
       if (err instanceof TotpError) {
-        sendTokenError(res, httpStatusForTokenError(err.code), err.code, err.message);
+        sendStoreError(res, httpStatusForStoreError(err.code), err.code, err.message);
       } else {
-        sendTokenError(res, 500, 'INTERNAL', message);
+        sendStoreError(res, 500, 'INTERNAL', message);
       }
     }
   });
